chore(zero-solid): rename useQuery to createQuery
The naming conventions for SolidJS is to name things createX over useX.
