+++ packages/zero-solid/src/mod.ts
@@ -1,2 +1,7 @@
+export {
+  createQuery,
+  useQuery,
+  type CreateQueryOptions,
+  type UseQueryOptions,
+} from './create-query.ts';
 export {createZero} from './create-zero.ts';
-export {useQuery, type UseQueryOptions} from './use-query.ts';
